requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.0.0
static3==0.5.1
wsgiref==0.1.2
django-filter==0.9.1
djangorestframework==3.0.2
markdown==2.5.2
pillow==2.7.0
